[REM] xoa bo code thuoc tinh va ham truong complete_name trong kderp_human_resources ( chuyen sang file kderp_extend_human_resources - trong mudule kderp_extend_module )
