+++ src/kderp_extend_module/__init__.py
@@ -33,7 +33,7 @@ import kderp_extend_contract
 #import kderp_extend_job_budget
 import kderp_extend_base
 import kderp_extend_supplier_payment
-
+import kderp_extend_human_resources
 import kderp_budget
 #import kderp_extend_other_expense
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
